requirements.txt
PyQt6>=6.4.0
PyQt6-Qt6>=6.4.0
Pillow>=10.0.0
rembg==2.0.56
numpy>=1.24.0
opencv-python>=4.8.0
onnxruntime>=1.16.0
